refactor: move api entities to shared domain
