+++ src/api/research-study/application/contracts/ResearchStudyRepository.ts
@@ -1,7 +1,7 @@
 import { Bundle } from 'fhir/r4'
 
+import { ElasticsearchBodyType } from '../../../../shared/elasticsearch/ElasticsearchBody'
 import { ResearchStudyQueryParams } from '../../controllers/converter/ResearchStudyQueryParams'
-import { ElasticsearchBodyType } from '../entities/ElasticsearchBody'
 
 export interface ResearchStudyRepository {
   findOne(id: string): Promise<unknown>
