+++ src/api/research-study/gateways/EsResearchStudyRepository.test.ts
@@ -3,8 +3,8 @@ import { expect } from 'vitest'
 import { EsResearchStudyRepository } from './EsResearchStudyRepository'
 import { EclaireDto } from '../../../etl/dto/EclaireDto'
 import { ResearchStudyModelFactory } from '../../../etl/factory/ResearchStudyModelFactory'
+import { ElasticsearchBodyType } from '../../../shared/elasticsearch/ElasticsearchBody'
 import { setupClientAndElasticsearchService } from '../../../shared/test/helpers/elasticsearchHelper'
-import { ElasticsearchBodyType } from '../application/entities/ElasticsearchBody'
 import { ResearchStudyQueryParams } from '../controllers/converter/ResearchStudyQueryParams'
 import { elasticsearchIndexMapping } from 'src/shared/elasticsearch/elasticsearchIndexMapping'
 import { SearchResponse } from 'src/shared/elasticsearch/ElasticsearchService'
