+++ src/shared/models/backbone-elements/BundleEntryModel.ts
