+++ src/shared/models/resources/BundleModel.ts
@@ -1,6 +1,6 @@
 import { Bundle, BundleEntry, BundleLink, FhirResource } from 'fhir/r4'
 
-import { BundleEntryModel } from './BundleEntryModel'
+import { BundleEntryModel } from '../backbone-elements/BundleEntryModel'
 import { SearchResponse, SearchResponseHits } from 'src/shared/elasticsearch/ElasticsearchService'
 
 export class BundleModel {
